+++ terraform/examples/identity-basic/outputs.tf
@@ -1,0 +1,17 @@
+output "role_name" {
+
+  value = module.identity.role_name
+
+}
+
+output "role_arn" {
+
+  value = module.identity.role_arn
+
+}
+
+output "policy_arn" {
+
+  value = module.identity.policy_arn
+
+}