+++ terraform/modules/identity/naming.tf
@@ -1,0 +1,11 @@
+locals {
+
+  role_name = lower(
+    "${local.resource_prefix}-${var.role_name}"
+  )
+
+  policy_name = lower(
+    "${local.resource_prefix}-${var.role_name}-policy"
+  )
+
+}